requirements.txt
aiogram>=3.0.0
mistralai
chromadb
aiosqlite
pydantic
python-dotenv